requirements.txt
flask==2.3.3
openai>=1.30.0
qdrant-client==1.6.0
PyPDF2==3.0.1
markdown==3.5.1
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
werkzeug==2.3.7

# Additional dependencies for enhanced functionality
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1

# Development dependencies
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
